src/main/java/BankAccount.java
public class BankAccount {

    private int balance;
    private String name;

    public BankAccount(int initialBalance, String name) {
        this.balance = initialBalance;
        this.name = name;
    }

    public int getBalance() {
        return balance;
    }

    public void deposit(int amount) {
        if (amount > 0) {
            balance += amount;
        } else {
            throw new IllegalArgumentException("Amount must be positive");
        }
    }

    public void withdraw(int amount) {
        if (amount > 0 && amount <= balance) {
            balance -= amount;
        }
    }
}
